+++ plat/fvp/Cargo.toml
@@ -26,6 +26,7 @@ islet_rmm = { path = "../../rmm" }
 linked_list_allocator = "0.10.4"
 log = "0.4.17"
 uart = { path = "../../lib/uart" }
+io = { path = "../../lib/io" }
 
 [build-dependencies]
 cc = "1.0"
