+++ rmm/Cargo.toml
@@ -23,6 +23,7 @@ spin = "0.9.2"
 spinning_top = "0.2.4"
 tinyvec = { version = "*", features = ["rustc_1_55"]}
 autopadding = { path = "../lib/autopadding" }
+io = { path = "../lib/io" }
 
 [build-dependencies]
 cc = "1.0"
